requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

gdown==5.2.0
huggingface_hub==0.30.1

ultralytics==8.3.163
opencv-python==4.11.0.86
numpy==2.0.2
scikit-learn==1.7.1

boxmot==15.0.2

librosa==0.11.0
soundfile==0.13.1
yt-dlp==2025.7.21

fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7

python-telegram-bot==22.3
